requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
imapclient>=3.0.0
pandas>=2.0.0
python-multipart>=0.0.6
reportlab>=4.0.0
pillow>=10.0.0
pytesseract>=0.3.10
python-docx>=1.1.0
openpyxl>=3.1.0
pdf2image>=1.16.3
pytest>=7.4.0
PyPDF2>=3.0.0
reportlab>=4.0.4
pytesseract>=0.3.10
python-docx>=1.1.0
openpyxl>=3.1.0
pdfplumber>=0.10.0  # Add this for better PDF extraction